requirements.txt
aioredis==1.0.0
amqp==2.2.2
anyjson==0.3.3
apns2==0.3.0
arrow==0.12.1
asgiref==2.1.6
asn1crypto==0.24.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
billiard==3.5.0.3
blessed==1.14.2
boto==2.48.0
boto3==1.6.16
botocore==1.9.16
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
channels==2.0.2
channels-redis==2.1.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
daphne==2.0.4
Django==2.0.2
django-picklefield==1.0.0
django-q==0.9.2
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
django-ses==0.8.5
djangorestframework==3.7.7
docutils==0.14
future==0.16.0
h2==2.6.2
hiredis==0.2.0
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
jmespath==0.9.3
kombu==4.1.0
mongoengine==0.15.0
msgpack==0.5.6
py-vapid==1.3.1
pycparser==2.18
PyJWT==1.5.3
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2018.3
pywebpush==1.7.0
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.47.0
zope.interface==4.4.3
